Update sigpyr script and dependencies

+++ sigpyr.sh
@@ -2,11 +2,53 @@
 
 TIMEOUT=./dep/timeout.sh
 
-#TODO: If I'm going to keep this script in this folder, I should include the current time in milliseconds in the filename, that way if the script is executed multiple times, the filenames wont clash.
-communicationFilename=".sigpyr_comm_file"
-nc -lvvdp 4444 > $communicationFilename &
+TIMEOUT_PERIOD_IN_SECONDS=5
+PORT_NUMBER=4444
+SIGNAL="ready"
 
-$TIMEOUT -t 5 ./dep/sigpyr_internal.sh
+POSITIONAL=()
+while [[ $# -gt 0 ]]
+do
+  key="$1"
+
+  case $key in
+    -p|--port)
+      PORT_NUMBER="$2"
+      shift
+      shift
+      ;;
+    -t|--timeout)
+      TIMEOUT_PERIOD_IN_SECONDS="$2"
+      shift
+      shift
+      ;;
+    -s|--signal)
+      SIGNAL="$2"
+      shift
+      shift
+      ;;
+    *)
+      POSITIONAL+=("$1")
+      shift
+      ;;
+  esac
+done
+
+set -- "{$POSITIONAL[@]}"
+
+if [ `netstat -ant | awk -v pattern="[^0-9]$PORT_NUMBER$" '$6 == "LISTEN" && $4 ~ pattern' | wc -l` != 0 ]
+then
+  echo "port $PORT_NUMBER is already in use"
+  exit 1
+fi
+
+# TODO: This should be generated in a method.
+MILLISECONDS_SINCE_EPOCH=$(( $(date '+%s%N') / 1000000));
+COMMUNICATION_FILENAME=".signpyr_comm_file_$MILLISECONDS_SINCE_EPOCH"
+
+nc -lvvdp $PORT_NUMBER > $COMMUNICATION_FILENAME &
+
+$TIMEOUT -t $TIMEOUT_PERIOD_IN_SECONDS ./dep/sigpyr_internal.sh $COMMUNICATION_FILENAME
 
 pkill -fx "nc -lvvdp 4444"
-rm -f $communicationFilename
+rm -f $COMMUNICATION_FILENAME
